requirements_prod.txt
Django==6.0.1
djangorestframework==3.14.0
django-cors-headers==4.3.1
django-crispy-forms==2.5
python-decouple==3.8
python-dotenv==1.2.2
Pillow==10.2.0
gunicorn==21.2.0
whitenoise==6.6.0
dj-database-url==2.1.0
psycopg2-binary==2.9.9
requests==2.32.5
razorpay==2.0.0
